rename and move optionRenderer
remove obselete import
