+++ src/components/commons/components/combo-box/panel/combo-box-option.spec.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import OptionRenderer from './option-renderer';
+import ComboBoxOption from './combo-box-option';
 import { describe, it, expect } from 'vitest';
 
 describe('DefaultOptionRenderer', () => {
 	it('renders without label', () => {
 		const option = { id: '1', value: 'Value' };
 		const { getByText, queryByText } = render(
-			<OptionRenderer option={option} />
+			<ComboBoxOption option={option} />
 		);
 		const idElement = getByText(option.id);
 		const labelElement = queryByText('-');
@@ -17,7 +17,7 @@ describe('DefaultOptionRenderer', () => {
 
 	it('renders with label', () => {
 		const option = { id: '1', value: 'Value', label: 'Label' };
-		const { getByText } = render(<OptionRenderer option={option} />);
+		const { getByText } = render(<ComboBoxOption option={option} />);
 		const idElement = getByText(option.id);
 		const labelElement = getByText(option.label);
 		expect(idElement).toBeInTheDocument();
@@ -26,13 +26,13 @@ describe('DefaultOptionRenderer', () => {
 
 	it('renders with selected class', () => {
 		const option = { id: '1', value: 'Value', label: 'Label' };
-		const { container } = render(<OptionRenderer option={option} selected />);
+		const { container } = render(<ComboBoxOption option={option} selected />);
 		expect(container.firstChild).toHaveClass('selected');
 	});
 
 	it('renders without selected class', () => {
 		const option = { id: '1', value: 'Value', label: 'Label' };
-		const { container } = render(<OptionRenderer option={option} />);
+		const { container } = render(<ComboBoxOption option={option} />);
 		expect(container.firstChild).not.toHaveClass('selected');
 	});
 });
