+++ src/components/commons/index.ts
@@ -1,4 +1,3 @@
-import { default as OptionRendererFn } from './components/option-renderer';
 export { default as LabelSelectionRenderer } from './components/label-selection-renderer';
 export { default as FieldContainer } from './components/field-container';
 export { default as Label } from './components/label';
@@ -12,6 +11,5 @@ export { default as createRowOrchestrator } from './create-row-orchestrator';
 export { default as OrchestratedComponent } from './components/orchestrated-component';
 export { default as ComboBox } from './components/combo-box';
 export * as HtmlTable from './components/html-table';
-export const OptionRenderer = OptionRendererFn; // We need to export this way to avoid crash on storybook TODO : investigate
 export { default as safetyLabel } from './safety-label';
 export { default as Errors } from './components/errors';
